Fix vuln policy table occasionally getting stuck in loading
The logic that is supposed to activate tabs based on the route (`this.getTabFromRoute()`) did not check the return value. This lead to fields / methods being called on an `undefined` object, which caused further loading of the page to crash. The vulnerability policy table was stuck in loading mode, even though the API returned a proper response.

Also fixed the vulnerability policy tab losing focus when reloading the page.